+++ src/migrations/m240530_122024_multisite_support.php
@@ -41,17 +41,19 @@ class m240530_122024_multisite_support extends Migration
         // Add new columns
         $this->_newColumns();
 
+        // Populate new columns with existing data
+        $this->_populateData();
+
+        // Make new column not nullable
+        //  - AFTER the column has been populated
+        //  - BEFORE adding indexes and foreign keys
+        $this->alterColumn(Install::GM_ADDRESSES, 'siteId', $this->integer()->notNull());
+
         // Add new column indexes
         $this->_newIndexes();
 
         // Add new foreign keys
         $this->_newForeignKeys();
-
-        // Populate new columns with existing data
-        $this->_populateData();
-
-        // Make new column not nullable, AFTER the column has been populated
-        $this->alterColumn(Install::GM_ADDRESSES, 'siteId', $this->integer()->notNull());
 
         // Post an announcement
         $this->_announcement();